refactor: #2 - rename server_settings_service to config_service, and re-organise cogs
